+++ include/GVCFReader.hpp
@@ -8,6 +8,7 @@
 #include <algorithm>
 #include <iostream>
 #include "utils.hpp"
+#include "numeric"
 
 extern "C" {
 #include <htslib/hts.h>
@@ -16,11 +17,37 @@ extern "C" {
 #include "vcfnorm.h"
 }
 
+//vcfnorm stuff
+#define ERR_DUP_ALLELE      -2
 #define ERR_REF_MISMATCH    -1
+#define ERR_OK              0
+#define ERR_SYMBOLIC        1
 #define CHECK_REF_WARN 1
-
 #define MROWS_SPLIT 1
 #define MROWS_MERGE  2
+//end vcf norm stuff
+
+int mnp_split(bcf1_t *record_to_split, bcf_hdr_t *header, vector<bcf1_t *> &output);
+
+class Genotype
+{
+public:
+    Genotype(bcf_hdr_t *header, bcf1_t *record);
+
+    Genotype(int ploidy, int num_allele);
+
+    Genotype marginalise(int index);
+
+    ~Genotype();
+
+    void setDepthFromAD();
+
+    int update_bcf1_t(bcf_hdr_t *header, bcf1_t *record);
+
+    int *_gt, *_ad, *_gq, *_dp, *_dpf, *_pl;
+    int _num_allele, _num_pl, _ploidy, _num_gt, _num_ad, _num_gq, _num_dp, _num_dpf, _num_gl;
+    std::vector<float> _gl;
+};
 
 //this basically wraps bcftools norm in a class.
 //TODO: investigate replacing this with invariant components