+++ src/test_gvcfgenotyper.cpp
@@ -6,14 +6,14 @@
 #define PATH_MAX 4096
 #endif
 
-GTestEnvironment* g_testenv = nullptr;
+GTestEnvironment *g_testenv = nullptr;
 
-int main(int argc, char** argv)
+int main(int argc, char **argv)
 {
     ::testing::InitGoogleTest(&argc, argv);
 
     // determine base path
-    char* bn = dirname(argv[0]);
+    char *bn = dirname(argv[0]);
     char actualpath[PATH_MAX + 1];
     realpath(bn, actualpath);