+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- heavily modified mnp_split function such that it will correctly handle (all?) multi-allelic MNPs
+- introduced the Genotype class which stores and manipulates our canonical FORMAT fields (currently GQ,DP,DPF,AD,PL)
 - buffers now store a fixed number of base pairs, not  number of variants
 - added -r argument to set regions
 - substantial performance improvements via better buffer memory management